Add files via upload
The zip file contains the necessary files for the Project2Data.R, which then writes the csv used in the Project2.Rmd. I am unable to attach the csv to the github repo as it is too large, so if that is needed for some reason let me know. Otherwise, take the files out of the zip file and place them in the working directory, which should allow the Project2Data.R file to execute and write the csv.